requirements.txt
pandas==2.2.2
sqlalchemy
psycopg2-binary
